removed unused code and refactor
