+++ app/src/main/java/com/ibm/rides/data/api/ApiService.kt
@@ -1,7 +1,7 @@
 package com.ibm.rides.data.api
 
 
-import com.ibm.rides.ui.vehicle.model.Vehicle
+import com.ibm.rides.domain.vehicle.list.Vehicle
 import com.ibm.rides.data.response.NetworkResponse
 import com.ibm.rides.utils.Constant
 import okhttp3.ResponseBody
