+++ app/src/main/java/com/ibm/rides/ui/vehicle/detail/VehicleDetailViewModel.kt
@@ -3,7 +3,7 @@ package com.ibm.rides.ui.vehicle.detail
 import android.util.Log
 import androidx.lifecycle.MutableLiveData
 import com.ibm.rides.basecontroller.BaseViewModel
-import com.ibm.rides.ui.vehicle.model.Vehicle
+import com.ibm.rides.domain.vehicle.list.Vehicle
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
