+++ app/src/test/java/com/ibm/rides/UserInputValidationUnitTest.kt
@@ -1,19 +1,35 @@
 package com.ibm.rides
 
 import com.ibm.rides.ui.vehicle.userinput.UserInputValidator
+import com.ibm.rides.ui.vehicle.userinput.UserInputViewModel
 import dagger.hilt.android.testing.HiltAndroidTest
 import org.junit.Assert
+import org.junit.Before
 import org.junit.Test
+import org.junit.runner.RunWith
+import org.mockito.Mock
+import org.mockito.junit.MockitoJUnitRunner
 import kotlin.random.Random
 
 @HiltAndroidTest
+@RunWith(MockitoJUnitRunner::class)
 class UserInputValidationUnitTest {
+
+
+    private lateinit var viewModel: UserInputViewModel
+    @Mock
+    private lateinit var validator: UserInputValidator
+
+    @Before
+    fun setup() {
+        viewModel = UserInputViewModel(validator)
+    }
 
     @Test
     fun `Assert user input validation`() {
         val expectedValidCount = true
 //        val isValidCount = UserInputValidator().isValidVehicleCount(Random.nextInt(101, 1000))
-        val isValidCount = UserInputValidator().isValidVehicleCount(Random.nextInt(1, 100))
+        val isValidCount = viewModel.isValidVehicleCount(Random.nextInt(1, 100))
         Assert.assertEquals(expectedValidCount, isValidCount)
     }
 }
